+++ src/triggers/commerce/OrderCompletedTrigger.php
@@ -27,4 +27,11 @@ class OrderCompletedTrigger extends BaseTrigger
         $order = $event->sender;
         return (float)$order->getTotalPrice();
     }
+
+    public static function getOrderIdFromEvent($event): ?int
+    {
+        /** @var Order $order */
+        $order = $event->sender;
+        return $order->id ?: null;
+    }
 }