+++ src/triggers/commerce/OrderPaidTrigger.php
@@ -28,4 +28,11 @@ class OrderPaidTrigger extends BaseTrigger
         $order = $event->sender;
         return (float) $order->getTotalPrice();
     }
+
+    public static function getOrderIdFromEvent($event): ?int
+    {
+        /** @var Order $order */
+        $order = $event->sender;
+        return $order->id ?: null;
+    }
 }